+++ packages/backend/src/services/chart/charts/federation.ts
@@ -2,6 +2,7 @@ import Chart, { KVs } from '../core.js';
 import { Followings, Instances } from '@/models/index.js';
 import { name, schema } from './entities/federation.js';
 import { fetchMeta } from '@/misc/fetch-meta.js';
+import { In, MoreThan, Not } from 'typeorm';
 
 /**
  * フェデレーションに関するチャート
@@ -28,7 +29,7 @@ export default class FederationChart extends Chart<typeof schema> {
 			.select('f.followerHost')
 			.where('f.followerHost IS NOT NULL');
 
-		const [sub, pub, pubsub] = await Promise.all([
+		const [sub, pub, pubsub, active] = await Promise.all([
 			Followings.createQueryBuilder('following')
 				.select('COUNT(DISTINCT following.followeeHost)')
 				.where('following.followeeHost IS NOT NULL')
@@ -52,12 +53,18 @@ export default class FederationChart extends Chart<typeof schema> {
 				.setParameters(pubsubSubQuery.getParameters())
 				.getRawOne()
 				.then(x => parseInt(x.count, 10)),
+			Instances.count({
+				host: Not(In(meta.blockedHosts)),
+				isSuspended: false,
+				lastCommunicatedAt: MoreThan(new Date(Date.now() - (1000 * 60 * 60 * 24 * 30))),
+			}),
 		]);
 
 		return {
 			'sub': sub,
 			'pub': pub,
 			'pubsub': pubsub,
+			'active': active,
 		};
 	}